fix: deprecate old signin flow
