+++ lang/cs.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Začíná v",
   "contribution.amount.currency.label": "Amount ({currency})",
   "contribution.amount.other.label": "Ostatní",
-  "contribution.createOrgProfile": "Vytvořit profil organizace",
-  "contribution.createPersoProfile": "Vytvořit osobní profil",
   "contribution.hostFeePercent": "Fiscal Host fee",
   "contribution.interval.label": "Frekvence",
   "contribution.itemPrice": "Cena produktu",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Vyberte profil",
   "createPledge.signinToCreate": "<signin-link>Sign in or join free</signin-link> to create a pledge.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "Máte již účet?",
   "createProfile.faq.anonymous.content": "Yes you can! Select the 'incognito' option to make your contribution anonymous to the public. However, the admins will still be able to see your profile and identity privately.",
   "createProfile.faq.anonymous.title": "Can I make an anonymous contribution?",
   "createProfile.faq.email.content": "The admins of the Collective and its Fiscal Host will have access to your email address. We do not share emails with anyone else and we don't use it for any type of marketing. We hate spam as much as you do.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "What's the difference between an individual and an organization profile?",
   "createProfile.faq.privacy.content": "We care about privacy. We don't use cookies, Google Analytics, or any kind of tracking. We collect certain personal information because it's required for regulatory reasons that the entity receiving money from you know who you are (KYC - Know Your Customer). You can choose to make an incognito contribution if you don't want your identity to be public.",
   "createProfile.faq.privacy.title": "A co soukromí?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "Informace o organizaci",
   "CreateProfile.PersonalInfo": "Vaše osobní informace",
   "createProject.create": "Vytvořit projekt",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Learn more",
   "LearnMoreAboutServiceFees": "Learn more about {service} fees",
   "LegalName": "Legal Name",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Share",
   "shareIt": "Sdílet",
   "signIn": "Přihlásit se",
-  "signin.joinForFree": "Připojte se zdarma!",
-  "signin.joinFree": "Připojit se zdarma",
-  "SignIn.legal": "By joining, you agree to our <TOSLink>Terms of Service</TOSLink> and <PrivacyPolicyLink>Privacy Policy</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Tvůj magický odkaz je na cestě!",
   "signin.noAccount": "Don't have an account?",
   "SignIn.SentTo": "We've sent it to {email}.",
   "SignIn.SuccessDetails": "You’ll be redirected from the link in the email, you can safely close this tab.",
-  "signin.unknownEmail": "There is no user with this email address.",
-  "signin.usingEmail": "Sign in using your email address:",
   "signinLinkSent.": "<Link>Learn more</Link> about passwordless login.",
   "SkipOnboarding": "Přeskočit průvodce",
   "somORZ": "Profile Type",
