+++ lang/de.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Beginnt bei",
   "contribution.amount.currency.label": "Betrag ({currency})",
   "contribution.amount.other.label": "Andere",
-  "contribution.createOrgProfile": "Erstelle ein Organisationsprofil",
-  "contribution.createPersoProfile": "Erstelle ein persönliches Profil",
   "contribution.hostFeePercent": "Steuerträger-Gebühr",
   "contribution.interval.label": "Häufigkeit",
   "contribution.itemPrice": "Artikelpreis",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Wähle ein Profil",
   "createPledge.signinToCreate": "<signin-link>Melden Sie sich an oder treten Sie kostenlos bei</signin-link>, um ein Versprechen zu erstellen.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "Haben Sie bereits ein Konto?",
   "createProfile.faq.anonymous.content": "Yes you can! Select the 'incognito' option to make your contribution anonymous to the public. However, the admins will still be able to see your profile and identity privately.",
   "createProfile.faq.anonymous.title": "Kann ich einen anonymen Beitrag leisten?",
   "createProfile.faq.email.content": "The admins of the Collective and its Fiscal Host will have access to your email address. We do not share emails with anyone else and we don't use it for any type of marketing. We hate spam as much as you do.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "What's the difference between an individual and an organization profile?",
   "createProfile.faq.privacy.content": "We care about privacy. We don't use cookies, Google Analytics, or any kind of tracking. We collect certain personal information because it's required for regulatory reasons that the entity receiving money from you know who you are (KYC - Know Your Customer). You can choose to make an incognito contribution if you don't want your identity to be public.",
   "createProfile.faq.privacy.title": "Was ist mit der Privatsphäre?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "Organisations-Info",
   "CreateProfile.PersonalInfo": "Deine personenbezogenen Daten",
   "createProject.create": "Erstelle ein Projekt",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Learn more",
   "LearnMoreAboutServiceFees": "Learn more about {service} fees",
   "LegalName": "Legal Name",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Teilen",
   "shareIt": "Share it",
   "signIn": "Anmelden",
-  "signin.joinForFree": "Kostenlos beitreten!",
-  "signin.joinFree": "Kostenlos beitreten",
-  "SignIn.legal": "By joining, you agree to our <TOSLink>Terms of Service</TOSLink> and <PrivacyPolicyLink>Privacy Policy</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Ihr magischer Link ist zu Ihnen unterwegs!",
   "signin.noAccount": "Sie haben noch kein Konto?",
   "SignIn.SentTo": "Wir haben es an {email} gesendet.",
   "SignIn.SuccessDetails": "Sie werden von dem Link in der E-Mail weitergeleitet, Sie können diesen Tab schließen.",
-  "signin.unknownEmail": "Es gibt kein Konto mit dieser E-Mail-Adresse.",
-  "signin.usingEmail": "Melden Sie sich mit Ihrer E-Mail-Adresse an:",
   "signinLinkSent.": "<Link>Learn more</Link> about passwordless login.",
   "SkipOnboarding": "Skip onboarding",
   "somORZ": "Profile Type",
