+++ lang/es.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Comienza desde",
   "contribution.amount.currency.label": "Cantidad ({currency})",
   "contribution.amount.other.label": "Otro",
-  "contribution.createOrgProfile": "Crear Perfil de Organización",
-  "contribution.createPersoProfile": "Crear Perfil Personal",
   "contribution.hostFeePercent": "Tarifa de Responsable Fiscal",
   "contribution.interval.label": "Frecuencia",
   "contribution.itemPrice": "Precio del articulo",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Escoge un perfil",
   "createPledge.signinToCreate": "<signin-link>Inicia sesión o únete gratis</signin-link> para crear un causa.",
   "createPledge.why": "Si la causa o colectivo que deseas apoyar aún no está en Open Collective, puedes plantear la iniciativa. Esto estimulará a sus líderes a crear un colectivo para sus actividades y le ofrecerá mucha más visibilidad sobre cómo se gasta su dinero para hacer avanzar su causa.",
-  "CreateProfile.AlreadyHaveAnAccount": "¿Ya tiene cuenta?",
   "createProfile.faq.anonymous.content": "¡Si, puedes hacerlo! Selecciona la opción \"Incógnito\" para hacer que tu contribución sea anónima. No obstante, los administradores podrán ver tu perfil y tu identidad de manera privada.",
   "createProfile.faq.anonymous.title": "¿Puedo hacer una aportación anónima?",
   "createProfile.faq.email.content": "Los administradores del Colectivo y su Anfitrión Fiscal tendrán acceso a su dirección de correo electrónico. No compartimos el correo electrónico con nadie más y no lo utilizamos para ningún tipo de marketing. Odiamos el spam tanto como tú.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "¿Cuál es la diferencia entre un perfil de individuo y un perfil de organización?",
   "createProfile.faq.privacy.content": "Nos importa la privacidad. No usamos cookies, Google Analytics ni ningún software de seguimiento. Recolectamos cierta información personal porque es necesaria debido a razones de regulación para que la entidad que está recibiendo tu dinero sepa quién eres (KYC - Know Your Customer; \"Conoce a tu cliente\"). Puedes elegir hacer una contribución anónima si no quieres que tu identidad sea pública.",
   "createProfile.faq.privacy.title": "¿Qué hay de la privacidad?",
-  "CreateProfile.name.conflict": "No puedes usar el mismo nombre para tu perfil personal y de organización. Los perfiles personales representan a personas individuales que pueden ser administradores de perfiles de organización.",
   "CreateProfile.OrgInfo": "Información de la organización",
   "CreateProfile.PersonalInfo": "Tu información personal",
   "createProject.create": "Crear un proyecto",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Leer más",
   "LearnMoreAboutServiceFees": "Leer más sobre las comisiones de {service}",
   "LegalName": "Nombre Legal",
-  "legalName.description": "El nombre legal es privado y se comparte con los anfitriones para los recibos de donaciones, los formularios de impuestos y cuando presentas un gasto. Este nombre no se muestra públicamente y debe ser tu nombre legal.",
   "lEZrUV": "¡Estamos contratando Ingenieros de Software Senior!<link><strong>Conozca más sobre los puestos vacantes</strong></link>",
   "LfwUBd": "Cuando contribuyes a un Colectivo, compartimos tu dirección de correo electrónico con los Administradores. Si prefieres mantener tu contribución en privado, elige el perfil de \"incógnito\". Lee nuestra <PrivacyPolicyLink>política de privacidad</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Compartir",
   "shareIt": "Compartirlo",
   "signIn": "Iniciar Sesión",
-  "signin.joinForFree": "¡Únete gratis!",
-  "signin.joinFree": "Unirte Gratis",
-  "SignIn.legal": "Al inscribirte, aceptas los <TOSLink>Términos de Servicio</TOSLink> y la <PrivacyPolicyLink>Política de Privacidad</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "¡Su enlace mágico está en camino!",
   "signin.noAccount": "¿No posee una cuenta?",
   "SignIn.SentTo": "Lo hemos enviado al {email}.",
   "SignIn.SuccessDetails": "Serás redirigido desde el enlace del correo electrónico, así que puedes cerrar esta pestaña sin problemas.",
-  "signin.unknownEmail": "No hay ningún usuario con esta dirección de correo electrónico.",
-  "signin.usingEmail": "Inicia sesión con tu dirección de correo electrónico:",
   "signinLinkSent.": "<Link>Más información</Link> sobre cómo iniciar sesión sin contraseña.",
   "SkipOnboarding": "Saltar procesos de integración",
   "somORZ": "Profile Type",
