+++ lang/fr.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "À partir de",
   "contribution.amount.currency.label": "Montant ({currency})",
   "contribution.amount.other.label": "Autre",
-  "contribution.createOrgProfile": "Créer un profil d'organisation",
-  "contribution.createPersoProfile": "Créer un profil personnel",
   "contribution.hostFeePercent": "Frais d'hôte fiscal",
   "contribution.interval.label": "Fréquence",
   "contribution.itemPrice": "Prix de l’article",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Choisissez un profil",
   "createPledge.signinToCreate": "<signin-link>Connectez-vous ou créez un compte</signin-link> pour effectuer une promesse de don.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "Vous avez déjà un compte ?",
   "createProfile.faq.anonymous.content": "Yes you can! Select the 'incognito' option to make your contribution anonymous to the public. However, the admins will still be able to see your profile and identity privately.",
   "createProfile.faq.anonymous.title": "Puis-je contribuer anonymement ?",
   "createProfile.faq.email.content": "The admins of the Collective and its Fiscal Host will have access to your email address. We do not share emails with anyone else and we don't use it for any type of marketing. We hate spam as much as you do.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "What's the difference between an individual and an organization profile?",
   "createProfile.faq.privacy.content": "We care about privacy. We don't use cookies, Google Analytics, or any kind of tracking. We collect certain personal information because it's required for regulatory reasons that the entity receiving money from you know who you are (KYC - Know Your Customer). You can choose to make an incognito contribution if you don't want your identity to be public.",
   "createProfile.faq.privacy.title": "Qu'en est-il de ma vie privée ?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "Organization info",
   "CreateProfile.PersonalInfo": "Vos informations personnelles",
   "createProject.create": "Créer un Projet",
@@ -1592,7 +1588,6 @@
   "LearnMore": "En apprendre plus",
   "LearnMoreAboutServiceFees": "En savoir plus sur les frais de {service}",
   "LegalName": "Nom Légal",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Partager",
   "shareIt": "Partager",
   "signIn": "S'identifier",
-  "signin.joinForFree": "Inscrivez-vous gratuitement !",
-  "signin.joinFree": "S'inscrire",
-  "SignIn.legal": "En continuant, vous acceptez nos <TOSLink>conditions d'utilisation</TOSLink> et notre <PrivacyPolicyLink>politique de confidentialité</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Votre lien magique est en route !",
   "signin.noAccount": "Pas encore de compte ?",
   "SignIn.SentTo": "Il a été envoyé à {email}.",
   "SignIn.SuccessDetails": "Vous serez redirigé·e depuis le lien dans l'email, vous pouvez fermer cet onglet.",
-  "signin.unknownEmail": "Il n'y a aucun utilisateur avec cette adresse e-mail.",
-  "signin.usingEmail": "Connectez-vous avec votre adresse email :",
   "signinLinkSent.": "<Link>En savoir plus</Link> sur la connexion sans mot de passe.",
   "SkipOnboarding": "Ignorer l'introduction",
   "somORZ": "Profile Type",
