+++ lang/ja.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "開始",
   "contribution.amount.currency.label": "額 ({currency})",
   "contribution.amount.other.label": "その他",
-  "contribution.createOrgProfile": "組織のプロフィールを作成",
-  "contribution.createPersoProfile": "個人プロフィールを作成",
   "contribution.hostFeePercent": "財政ホスト手数料",
   "contribution.interval.label": "頻度",
   "contribution.itemPrice": "料金",
@@ -757,7 +755,6 @@
   "createPledge.profile": "プロフィールを選択",
   "createPledge.signinToCreate": "<signin-link>Sign in or join free</signin-link> to create a pledge.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "既にアカウントをお持ちですか?",
   "createProfile.faq.anonymous.content": "可能です! 匿名オプションを選択することで貢献を匿名にして公開できます。しかし、管理者は秘密でプロフィールと身元を見ることができます。",
   "createProfile.faq.anonymous.title": "匿名で貢献することはできますか?",
   "createProfile.faq.email.content": "The admins of the Collective and its Fiscal Host will have access to your email address. We do not share emails with anyone else and we don't use it for any type of marketing. We hate spam as much as you do.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "What's the difference between an individual and an organization profile?",
   "createProfile.faq.privacy.content": "We care about privacy. We don't use cookies, Google Analytics, or any kind of tracking. We collect certain personal information because it's required for regulatory reasons that the entity receiving money from you know who you are (KYC - Know Your Customer). You can choose to make an incognito contribution if you don't want your identity to be public.",
   "createProfile.faq.privacy.title": "プライバシーはどうなりますか?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "組織の情報",
   "CreateProfile.PersonalInfo": "個人情報",
   "createProject.create": "Create a Project",
@@ -1592,7 +1588,6 @@
   "LearnMore": "詳細",
   "LearnMoreAboutServiceFees": "{service} 手数料の詳細",
   "LegalName": "法的な名前",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Share",
   "shareIt": "共有",
   "signIn": "サインイン",
-  "signin.joinForFree": "無料で登録!",
-  "signin.joinFree": "無料で登録",
-  "SignIn.legal": "参加することで、<TOSLink>利用規約</TOSLink>と<PrivacyPolicyLink>プライバシーポリシー</PrivacyPolicyLink>に同意したことになります。",
   "SignIn.LinkSent": "ログイン用リンクを送信しました！",
   "signin.noAccount": "アカウントを持っていませんか?",
   "SignIn.SentTo": "{email} にメールを送信しました。",
   "SignIn.SuccessDetails": "メール内のリンクからリダイレクトされるので、このタブを安全に閉じることができます。",
-  "signin.unknownEmail": "このメールアドレスを使っているユーザーはいません。",
-  "signin.usingEmail": "メールアドレスでログイン:",
   "signinLinkSent.": "パスワードレスログインについての詳細は<Link>こちら</Link>をご覧ください。",
   "SkipOnboarding": "Skip onboarding",
   "somORZ": "Profile Type",
