+++ lang/pl.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Rozpocznij od",
   "contribution.amount.currency.label": "Kwota ({currency})",
   "contribution.amount.other.label": "Inne",
-  "contribution.createOrgProfile": "Utwórz profil organizacji",
-  "contribution.createPersoProfile": "Utwórz profil osobisty",
   "contribution.hostFeePercent": "Opłata dla gospodarza podatkowego",
   "contribution.interval.label": "Częstotliwość",
   "contribution.itemPrice": "Cena produktu",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Wybierz profil",
   "createPledge.signinToCreate": "<signin-link>Zaloguj się lub dołącz za darmo</signin-link> aby utworzyć przyrzeczenie.",
   "createPledge.why": "Jeśli sprawa lub zbiórka, który chcesz wesprzeć, nie znajduje się jeszcze w Open Collective, możesz złożyć przysięgę. To zmotywuje do stworzenia otwartej zbiórki dla ich działań oraz zapewni znacznie lepszy wgląd, w jaki sposób wydawane są twoje pieniądze na rozwój ich sprawy.",
-  "CreateProfile.AlreadyHaveAnAccount": "Masz już konto?",
   "createProfile.faq.anonymous.content": "Tak! Wybierz opcję 'incognito', aby twój wkład był anonimowy dla społeczeństwa. Administratorzy będą jednak nadal mogli zobaczyć Twój profil i tożsamość prywatnie.",
   "createProfile.faq.anonymous.title": "Czy mogę dokonać anonimowej wpłaty?",
   "createProfile.faq.email.content": "Administratorzy Zbiórki i jego Gospodarz podatkowy będą mieli dostęp do Twojego adresu e-mail. Nie udostępniamy tego adresu nikomu innemu i nie używamy go do celów marketingowych. Nienawidzimy spamu tak samo jak Ty.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "Jaka jest różnica między profilem indywidualnym a profilem organizacji?",
   "createProfile.faq.privacy.content": "Dbamy o prywatność. Nie używamy plików cookie, Google Analytics, ani żadnego rodzaju śledzenia. Zbieramy pewne dane osobowe, ponieważ ze względów regulacyjnych podmiot otrzymujący pieniądze od Ciebie wie, kim jesteś (KYC - Know Your Customer). Możesz wybrać wkład anonimowy, jeśli nie chcesz, aby Twoja tożsamość była publiczna.",
   "createProfile.faq.privacy.title": "A co z prywatnością?",
-  "CreateProfile.name.conflict": "Nie możesz używać tej samej nazwy dla swoich profili indywidualnych i organizacyjnych. Profile indywidualne reprezentują poszczególne osoby, które mogą być administratorami profili Organizacji.",
   "CreateProfile.OrgInfo": "Organizacja",
   "CreateProfile.PersonalInfo": "Twoje informacje osobiste",
   "createProject.create": "Utwórz projekt",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Dowiedz się więcej",
   "LearnMoreAboutServiceFees": "Dowiedz się więcej o opłatach {service}",
   "LegalName": "Nazwa prawna",
-  "legalName.description": "Nazwa prawna jest prywatna i jest udostępniana gospodarzom na potrzeby wpływów darowizn, formularzy podatkowych i kiedy zgłaszasz wydatki. Ta nazwa nie jest wyświetlana publicznie i musi to być twoją prawdziwą nazwą jak np. imię i nazwisko.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "Kiedy wnosisz wkład do Zbiórki, udostępniamy Twój adres e-mail Administratorom. Jeśli chcesz zachować prywatność swojego wkładu, wybierz profil „incognito”. Przeczytaj naszą <PrivacyPolicyLink>politykę prywatności</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Udostępnij",
   "shareIt": "Udostępnij",
   "signIn": "Zaloguj się",
-  "signin.joinForFree": "Dołącz za darmo!",
-  "signin.joinFree": "Dołącz za darmo",
-  "SignIn.legal": "Dołączając do nas, zgadzasz się na nasze <TOSLink>Warunki świadczenia usług</TOSLink> oraz na <PrivacyPolicyLink>Politykę prywatności</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Twój magiczny link jest w drodze!",
   "signin.noAccount": "Nie posiadasz konta?",
   "SignIn.SentTo": "Wysłaliśmy na adres {email}.",
   "SignIn.SuccessDetails": "Zostaniesz przekierowany z linku w wiadomości e-mail, możesz bezpiecznie zamknąć tę kartę.",
-  "signin.unknownEmail": "Nie ma użytkownika z tym adresem e-mail.",
-  "signin.usingEmail": "Zaloguj się przy użyciu adresu e-mail:",
   "signinLinkSent.": "<Link>Dowiedz się więcej</Link> o logowaniu bez hasła.",
   "SkipOnboarding": "Skip onboarding",
   "somORZ": "Profile Type",
