+++ lang/pt-BR.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Começa em",
   "contribution.amount.currency.label": "Quantia ({currency})",
   "contribution.amount.other.label": "Outro",
-  "contribution.createOrgProfile": "Crie um perfil para a organização",
-  "contribution.createPersoProfile": "Criar perfil pessoal",
   "contribution.hostFeePercent": "Taxa de administrador fiscal",
   "contribution.interval.label": "Frequência",
   "contribution.itemPrice": "Preço do item",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Escolher um perfil",
   "createPledge.signinToCreate": "<signin-link>Entre ou participe de graça</signin-link> para se comprometer.",
   "createPledge.why": "Se a causa ou o coletivo que você deseja apoiar ainda não está na Open Collective, você pode fazer uma comprometimento. Isso irá incentivá-los a criar um coletivo aberto para suas atividades e oferecer à você muito mais visibilidade sobre como seu dinheiro é gasto para promover as causas deles.",
-  "CreateProfile.AlreadyHaveAnAccount": "Já possui uma conta?",
   "createProfile.faq.anonymous.content": "Sim, você pode! Selecione a opção \"incógnito\" para tornar sua contribuição anônima ao público. No entanto, os administradores ainda poderão ver seu perfil e sua identidade de forma privada.",
   "createProfile.faq.anonymous.title": "Eu posso fazer uma contribuição anônima?",
   "createProfile.faq.email.content": "Os administradores do coletivo e seu gestor de administração fiscal terão acesso ao seu endereço de e-mail. Nós não compartilhamos e-mails com mais ninguém e não os usamos para qualquer tipo de ação de marketing. Nós odiamos spam tanto quanto você.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "Qual é a diferença entre um perfil de indivíduo e um de organização?",
   "createProfile.faq.privacy.content": "Nós nos preocupamos com a privacidade. Não usamos cookies, Google Analytics, ou qualquer tipo de rastreamento. Nós coletamos certos dados pessoais por ser necessário, devido a motivos regulatórios, que a entidade recebendo dinheiro de você sabe quem você é (KYC - Conheça seu Cliente em português). Você pode optar por fazer uma contribuição anônima se não quiser que sua identidade seja pública.",
   "createProfile.faq.privacy.title": "E quanto à privacidade?",
-  "CreateProfile.name.conflict": "Você não pode usar o mesmo nome para os seus perfis pessoais e de organização. Perfis pessoais representam pessoas individuais, que podem ser administradores de perfis de organização.",
   "CreateProfile.OrgInfo": "Dados da organização",
   "CreateProfile.PersonalInfo": "Suas informações pessoais",
   "createProject.create": "Criar um Projeto",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Saiba mais",
   "LearnMoreAboutServiceFees": "Saiba mais sobre {service} tarifas",
   "LegalName": "Legal Name",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Compartilhar",
   "shareIt": "Compartilhar isto",
   "signIn": "Entrar",
-  "signin.joinForFree": "Junte-se de graça!",
-  "signin.joinFree": "Entrar de Graça",
-  "SignIn.legal": "Ao aderir, você concorda com nossos <TOSLink>Termos de Serviço</TOSLink> e a <PrivacyPolicyLink>Política de Privacidade</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Seu link mágico está a caminho!",
   "signin.noAccount": "Não tem uma conta?",
   "SignIn.SentTo": "Nós enviamos para {email}.",
   "SignIn.SuccessDetails": "Você será redirecionado a partir do link no e-mail, pode fechar esta aba com segurança.",
-  "signin.unknownEmail": "Não há usuário com este endereço de e-mail.",
-  "signin.usingEmail": "Entre usando seu endereço de e-mail:",
   "signinLinkSent.": "<Link>Saiba mais</Link> sobre login sem senha.",
   "SkipOnboarding": "Pular integração",
   "somORZ": "Profile Type",
