+++ lang/ru.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Начиная с",
   "contribution.amount.currency.label": "Сумма ({currency})",
   "contribution.amount.other.label": "Другое",
-  "contribution.createOrgProfile": "Создать профиль организации",
-  "contribution.createPersoProfile": "Создать личный профиль",
   "contribution.hostFeePercent": "Fiscal Host fee",
   "contribution.interval.label": "Частота",
   "contribution.itemPrice": "Цена",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Выбрать профиль",
   "createPledge.signinToCreate": "<signin-link>Sign in or join free</signin-link> to create a pledge.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "У вас уже есть учетная запись?",
   "createProfile.faq.anonymous.content": "Да, ты можешь! Выберите опцию \"инкогнито\", чтобы сделать ваш вклад анонимным для публики. Однако администраторы в частном порядке по-прежнему смогут видеть ваш профиль и личность.",
   "createProfile.faq.anonymous.title": "Могу ли я сделать анонимный взнос?",
   "createProfile.faq.email.content": "Администраторы Коллектива и его Финансового хоста будут иметь доступ к вашему адресу электронной почты. Мы никому не передаем электронные письма и не используем их для какого-либо маркетинга. Мы ненавидим спам также как и вы.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "В чем разница между индивидуальным профилем и профилем организации?",
   "createProfile.faq.privacy.content": "Мы заботимся о конфиденциальности. Мы не используем файлы cookie, Google Analytics или какие-либо средства отслеживания. Мы собираем определенную личную информацию, потому что по нормативным причинам требуется, чтобы организация, получающая от вас деньги, знала, кто вы (KYC — знай своего клиента). Вы можете сделать вклад инкогнито, если не хотите, чтобы ваша личность была общедоступной.",
   "createProfile.faq.privacy.title": "Как насчет конфиденциальности?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "Информация об организации",
   "CreateProfile.PersonalInfo": "Ваша личная информация",
   "createProject.create": "Создать Проект",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Узнать больше",
   "LearnMoreAboutServiceFees": "Подробнее о комиссии {service}",
   "LegalName": "Официальное имя",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Поделиться",
   "shareIt": "Поделиться",
   "signIn": "Вход",
-  "signin.joinForFree": "Присоединяйтесь бесплатно!",
-  "signin.joinFree": "Присоединиться бесплатно",
-  "SignIn.legal": "Зарегистрировавшись, вы соглашаетесь с <TOSLink>Условиями предоставления услуг</TOSLink> и <PrivacyPolicyLink>Политикой конфиденциальности</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Ваша волшебная ссылка уже в пути!",
   "signin.noAccount": "Еще нет аккаунта?",
   "SignIn.SentTo": "Мы отправили его на {email}.",
   "SignIn.SuccessDetails": "Вы будете перенаправлены по ссылке в письме. Вы можете закрыть эту вкладку.",
-  "signin.unknownEmail": "Нет пользователя с такой почтой.",
-  "signin.usingEmail": "Авторизоваться по электронной почте:",
   "signinLinkSent.": "<Link>Узнайте больше</Link> об аутентификации без пароля.",
   "SkipOnboarding": "Пропустить стартовый экран",
   "somORZ": "Profile Type",
