+++ lang/uk.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "Починається з",
   "contribution.amount.currency.label": "Сума ({currency})",
   "contribution.amount.other.label": "Інше",
-  "contribution.createOrgProfile": "Створити профіль організації",
-  "contribution.createPersoProfile": "Створити особовий профіль",
   "contribution.hostFeePercent": "Комісія фіскального вузла",
   "contribution.interval.label": "Частота",
   "contribution.itemPrice": "Ціна позиції",
@@ -757,7 +755,6 @@
   "createPledge.profile": "Вибрати профіль",
   "createPledge.signinToCreate": "<signin-link>Увійдіть або приєднайтеся безплатно</signin-link>, щоб створити зобов'язання.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "Вже маєте обліковий запис?",
   "createProfile.faq.anonymous.content": "Так, можете! Виберіть параметр «анонімно», щоб зробити ваш внесок анонімним для громадськості. Проте, адміністратори все одно приватно зможуть бачити ваш профіль і особу.",
   "createProfile.faq.anonymous.title": "Чи можна зробити анонімний внесок?",
   "createProfile.faq.email.content": "Адміністратори Колективу та його фіскальний вузол матимуть доступ до вашої адреси електронної пошти. Ми не ділимося адресами електронної пошти з будь-ким іншим і не користуємося ними для торгівлі. Ми ненавидимо спам, так само як і ви.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "У чому полягає різниця між особистим та профілем організації?",
   "createProfile.faq.privacy.content": "Ми піклуємося про приватність, ми не застосовуємо файли cookie, Google Analytics, або будь-який вид відстеження. Ми збираємо деякі особисті дані, тому що це потрібно з регуляторних причин, щоб отримувач коштів від вас знав, хто ви (KYC — знайте свого клієнта). Ви можете зробити внесок анонімно, якщо ви не хочете, щоб ваша особа була відома всім.",
   "createProfile.faq.privacy.title": "А як щодо приватності?",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "Відомості про організацію",
   "CreateProfile.PersonalInfo": "Ваші особисті дані",
   "createProject.create": "Створити проєкт",
@@ -1592,7 +1588,6 @@
   "LearnMore": "Докладніше",
   "LearnMoreAboutServiceFees": "Докладніше про платню для {service}",
   "LegalName": "Юридично точне ім'я",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "Поділитися",
   "shareIt": "Поділитись",
   "signIn": "Увійти",
-  "signin.joinForFree": "Приєднуйтеся безплатно!",
-  "signin.joinFree": "Приєднатися безплатно",
-  "SignIn.legal": "Приєднавшись, ви погоджуєтеся з нашими <TOSLink>Умовами користування</TOSLink> та <PrivacyPolicyLink>Політикою приватності</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "Ваше чарівне посилання вже в дорозі!",
   "signin.noAccount": "Ще не зареєстровані?",
   "SignIn.SentTo": "Ми надіслали його на {email}.",
   "SignIn.SuccessDetails": "You’ll be redirected from the link in the email, you can safely close this tab.",
-  "signin.unknownEmail": "Немає жодного користувача з цією електронною адресою.",
-  "signin.usingEmail": "Увійти за допомогою адреси електронної пошти:",
   "signinLinkSent.": "<Link>Докладніше</Link> про безпарольний вхід.",
   "SkipOnboarding": "Пропустити початкові настанови",
   "somORZ": "Profile Type",
