+++ lang/zh.json
@@ -594,8 +594,6 @@
   "ContributeTier.StartsAt": "起步",
   "contribution.amount.currency.label": "总额（{currency}）",
   "contribution.amount.other.label": "其他",
-  "contribution.createOrgProfile": "建立组织资料",
-  "contribution.createPersoProfile": "创建个人资料",
   "contribution.hostFeePercent": "Fiscal Host fee",
   "contribution.interval.label": "频率",
   "contribution.itemPrice": "Item price",
@@ -757,7 +755,6 @@
   "createPledge.profile": "选择简介",
   "createPledge.signinToCreate": "<signin-link>Sign in or join free</signin-link> to create a pledge.",
   "createPledge.why": "If the cause or collective that you want to support is not yet on Open Collective, you can make a pledge. This will incentivize them to create an open collective for their activities and offer you much more visibility on how your money is spent to advance their cause.",
-  "CreateProfile.AlreadyHaveAnAccount": "已有账户？",
   "createProfile.faq.anonymous.content": "是的，您可以！选择“隐身”选项来使您的贡献对公众匿名。然而，管理员仍然能够在私下看到您的资料和身份。",
   "createProfile.faq.anonymous.title": "我可以匿名贡献吗？",
   "createProfile.faq.email.content": "The admins of the Collective and its Fiscal Host will have access to your email address. We do not share emails with anyone else and we don't use it for any type of marketing. We hate spam as much as you do.",
@@ -768,7 +765,6 @@
   "createProfile.faq.persoVSOrg.title": "个人与组织资料之间有什么区别？",
   "createProfile.faq.privacy.content": "We care about privacy. We don't use cookies, Google Analytics, or any kind of tracking. We collect certain personal information because it's required for regulatory reasons that the entity receiving money from you know who you are (KYC - Know Your Customer). You can choose to make an incognito contribution if you don't want your identity to be public.",
   "createProfile.faq.privacy.title": "有关隐私的信息？",
-  "CreateProfile.name.conflict": "You can't use the same name for your Personal and Organization profiles. Personal profiles represent individual people, who can be administrators of Organization profiles.",
   "CreateProfile.OrgInfo": "组织信息",
   "CreateProfile.PersonalInfo": "您的个人信息",
   "createProject.create": "创建一个项目",
@@ -1592,7 +1588,6 @@
   "LearnMore": "了解更多",
   "LearnMoreAboutServiceFees": "进一步了解 {service} 费用",
   "LegalName": "Legal Name",
-  "legalName.description": "The legal name is private and shared with the hosts for donation receipts, tax forms and when you submit an expense. This name is not displayed publicly and it must be your legal name.",
   "lEZrUV": "We're hiring Senior Software Engineers! <link><strong>Learn more about open positions</strong></link>",
   "LfwUBd": "When you contribute to a Collective we share your email address with the Administrators. If you wish to keep your contribution private choose the ‘incognito’ profile. Read our <PrivacyPolicyLink>privacy policy</PrivacyPolicyLink>.",
   "Lg6nmh": "This collective doesn’t satisfy the minimum admin requirements as admin invitations are still pending.",
@@ -2144,15 +2139,10 @@
   "Share": "分享",
   "shareIt": "分享这个",
   "signIn": "登录",
-  "signin.joinForFree": "免费加入！",
-  "signin.joinFree": "免费注册",
-  "SignIn.legal": "By joining, you agree to our <TOSLink>Terms of Service</TOSLink> and <PrivacyPolicyLink>Privacy Policy</PrivacyPolicyLink>.",
   "SignIn.LinkSent": "你的魔法链接在路上了！",
   "signin.noAccount": "没有账户？",
   "SignIn.SentTo": "We've sent it to {email}.",
   "SignIn.SuccessDetails": "You’ll be redirected from the link in the email, you can safely close this tab.",
-  "signin.unknownEmail": "There is no user with this email address.",
-  "signin.usingEmail": "使用你的电子邮件注册：",
   "signinLinkSent.": "<Link>Learn more</Link> about passwordless login.",
   "SkipOnboarding": "跳过入职",
   "somORZ": "Profile Type",
