--- pages/signinv2.js
@@ -1,238 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import { mapValues } from 'lodash';
-import { withRouter } from 'next/router';
-import { FormattedMessage } from 'react-intl';
-import { isEmail } from 'validator';
-
-import { getFromLocalStorage, LOCAL_STORAGE_KEYS } from '../lib/local-storage';
-import { isSuspiciousUserAgent, RobotsDetector } from '../lib/robots-detector';
-import { isValidRelativeUrl } from '../lib/utils';
-
-import Body from '../components/Body';
-import { Flex } from '../components/Grid';
-import Header from '../components/Header';
-import Loading from '../components/Loading';
-import LoadingGrid from '../components/LoadingGrid';
-import MessageBox from '../components/MessageBox';
-import SignInOrJoinFreeV2 from '../components/SignInOrJoinFreeV2';
-import { P } from '../components/Text';
-import { withUser } from '../components/UserProvider';
-
-class SigninV2Page extends React.Component {
-  static getInitialProps({ query: { token, next, form, email }, req }) {
-    // Decode next URL if URI encoded
-    if (next && next.startsWith('%2F')) {
-      next = decodeURIComponent(next);
-    }
-
-    next = next && isValidRelativeUrl(next) ? next : null;
-    email = email && decodeURIComponent(email);
-    return {
-      token,
-      next,
-      form: form || 'signinv2',
-      isSuspiciousUserAgent: isSuspiciousUserAgent(req?.get('User-Agent')),
-      email: email && isEmail(email) ? email : null,
-    };
-  }
-
-  static propTypes = {
-    form: PropTypes.oneOf(['signinv2', 'create-accountv2']).isRequired,
-    token: PropTypes.string,
-    email: PropTypes.string,
-    next: PropTypes.string,
-    login: PropTypes.func,
-    errorLoggedInUser: PropTypes.string,
-    LoggedInUser: PropTypes.object,
-    loadingLoggedInUser: PropTypes.bool,
-    enforceTwoFactorAuthForLoggedInUser: PropTypes.bool,
-    isSuspiciousUserAgent: PropTypes.bool,
-    router: PropTypes.object,
-  };
-
-  constructor(props) {
-    super(props);
-    this.robotsDetector = new RobotsDetector();
-    this.state = { error: null, success: null, isRobot: props.isSuspiciousUserAgent, redirecting: false };
-  }
-
-  componentDidMount() {
-    if (this.state.isRobot) {
-      this.robotsDetector.startListening(() => this.setState({ isRobot: false }));
-    } else {
-      this.initialize();
-    }
-  }
-
-  async componentDidUpdate(oldProps, oldState) {
-    const wasConnected = !oldProps.LoggedInUser && this.props.LoggedInUser;
-
-    if (oldState.isRobot && !this.state.isRobot) {
-      this.initialize();
-    } else if (wasConnected && !this.props.errorLoggedInUser && this.props.form !== 'create-accountv2') {
-      // --- User logged in ---
-      this.setState({ success: true, redirecting: true });
-      // Avoid redirect loop: replace '/signinv2' redirects by '/'
-      const { next } = this.props;
-      const redirect = next && next.match(/^\/?signinv2[?/]?/) ? null : next;
-      await this.props.router.replace(redirect || '/');
-      window.scroll(0, 0);
-    } else if (this.props.token && oldProps.token !== this.props.token) {
-      // --- There's a new token in town 🤠 ---
-      const user = await this.props.login(this.props.token);
-      if (!user) {
-        this.setState({ error: 'Token rejected' });
-      }
-    }
-  }
-
-  componentWillUnmount() {
-    this.robotsDetector.stopListening();
-  }
-
-  async initialize() {
-    if (this.props.token) {
-      let user;
-      try {
-        user = await this.props.login(this.props.token);
-
-        // If given token is invalid, try to login with the old one
-        if (!user) {
-          user = await this.props.login();
-        }
-
-        // If there's no user at this point, there's no chance we can login
-        if (!user) {
-          this.setState({ error: 'Token rejected' });
-        }
-      } catch (err) {
-        this.setState({ error: err.message || err });
-      }
-    } else {
-      this.props.login();
-    }
-  }
-
-  getRoutes() {
-    const { next } = this.props;
-    const routes = { signin: '/signinv2', join: '/create-accountv2' };
-    if (!next) {
-      return routes;
-    } else {
-      const urlParams = `?next=${encodeURIComponent(next)}`;
-      return mapValues(routes, route => `${route}${urlParams}`);
-    }
-  }
-
-  renderContent() {
-    const {
-      loadingLoggedInUser,
-      errorLoggedInUser,
-      token,
-      next,
-      form,
-      LoggedInUser,
-      enforceTwoFactorAuthForLoggedInUser,
-    } = this.props;
-
-    if (this.state.isRobot && token) {
-      return (
-        <Flex flexDirection="column" alignItems="center" px={3} pb={3}>
-          <P fontSize="30px" mb={3}>
-            <span role="img" aria-label="Robot Emoji">
-              🤖
-            </span>
-          </P>
-          <P mb={5} textAlign="center">
-            <FormattedMessage
-              id="checkingBrowser"
-              defaultMessage="Your browser is being verified. If this message doesn't disappear, try to move your mouse or to touch your screen for mobile."
-            />
-          </P>
-          <Loading />
-        </Flex>
-      );
-    } else if ((loadingLoggedInUser || this.state.success) && token) {
-      return <Loading />;
-    } else if (!loadingLoggedInUser && LoggedInUser && form === 'create-accountv2') {
-      return (
-        <MessageBox type="warning" withIcon>
-          <FormattedMessage
-            id="createAccount.alreadyLoggedIn"
-            defaultMessage={`It seems like you're already signed in as "{email}". If you want to create a new account, please log out first.`}
-            values={{ email: LoggedInUser.email }}
-          />
-        </MessageBox>
-      );
-    }
-
-    const error = errorLoggedInUser || this.state.error;
-
-    if (loadingLoggedInUser || this.state.redirecting) {
-      return <LoadingGrid />;
-    }
-
-    return (
-      <React.Fragment>
-        {error && !error.includes('Two-factor authentication is enabled') && (
-          <MessageBox type="error" withIcon mb={4} data-cy="signin-message-box">
-            <strong>
-              <FormattedMessage
-                id="login.failed"
-                defaultMessage="Sign In failed: {message}."
-                values={{ message: error }}
-              />
-            </strong>
-            <br />
-            {!error?.includes('Two-factor authentication') && (
-              <FormattedMessage
-                id="login.askAnother"
-                defaultMessage="You can ask for a new sign in link using the form below."
-              />
-            )}
-          </MessageBox>
-        )}
-        <SignInOrJoinFreeV2
-          email={this.props.email}
-          redirect={next || '/'}
-          form={form}
-          routes={this.getRoutes()}
-          enforceTwoFactorAuthForLoggedInUser={enforceTwoFactorAuthForLoggedInUser}
-          submitTwoFactorAuthenticatorCode={twoFactorAuthenticatorCode => {
-            const localStorage2FAToken = getFromLocalStorage(LOCAL_STORAGE_KEYS.ACCESS_TOKEN);
-            return this.props.login(localStorage2FAToken, {
-              twoFactorAuthenticatorCode,
-            });
-          }}
-          submitRecoveryCode={recoveryCode => {
-            const localStorage2FAToken = getFromLocalStorage(LOCAL_STORAGE_KEYS.ACCESS_TOKEN);
-            return this.props.login(localStorage2FAToken, { recoveryCode });
-          }}
-        />
-      </React.Fragment>
-    );
-  }
-
-  render() {
-    return (
-      <div className="LoginPage">
-        <Header
-          title={this.props.form === 'signinv2' ? 'Sign In' : 'Create Account'}
-          description="Create your profile on Open Collective and show the world the open collectives that you are contributing to."
-          menuItemsV2={{ solutions: false, product: false, company: false, docs: false }}
-          menuItems={{ discover: false, docs: false, howItWorks: false, pricing: false }}
-          showSearch={false}
-          showProfileAndChangelogMenu={false}
-        />
-        <Body>
-          <Flex flexDirection="column" alignItems="center" my={[4, 6]} p={2}>
-            {this.renderContent()}
-          </Flex>
-        </Body>
-      </div>
-    );
-  }
-}
-
-export default withUser(withRouter(SigninV2Page));
