--- stories/design-system/SignInV2.stories.mdx
@@ -1,59 +1,0 @@
-import { Meta, Story, Canvas } from '@storybook/addon-docs/blocks';
-import SignInV2 from '../../components/SignInV2';
-
-<Meta title="Design system/SignInV2" component={SignInV2} />
-
-"Sign In" button stays disabled until there is a value in the email input.
-
-Uses the built-in browser email validation to block submission until valid.
-
-### Base form
-
-<Canvas>
-  <Story name="Base Form">
-    {() => {
-      const [action, setAction] = React.useState(null);
-      const [email, setEmail] = React.useState('');
-      return (
-        <div>
-          <SignInV2
-            onSubmit={value => setAction(`The form was submitted with a value of ${value}`)}
-            onSecondaryAction={() => setAction('Secondary action was clicked')}
-            onEmailChange={setEmail}
-            email={email}
-          />
-          <p>{action}</p>
-        </div>
-      );
-    }}
-  </Story>
-</Canvas>
-
-### Loading / submitting
-
-<Canvas>
-  <Story name="Loading / Submitting">
-    {() => {
-      return (
-        <SignInV2 loading onSubmit={() => {}} onSecondaryAction={() => {}} email="test+betree@opencollective.com" />
-      );
-    }}
-  </Story>
-</Canvas>
-
-### With an unknown email
-
-<Canvas>
-  <Story name="With Unknown Email">
-    {() => {
-      return (
-        <SignInV2
-          onSubmit={() => {}}
-          onSecondaryAction={() => {}}
-          unknownEmail
-          email="test+betree@opencollective.com"
-        />
-      );
-    }}
-  </Story>
-</Canvas>
